heroku

+++ client/apps/city_issues/models/issues.py
@@ -42,7 +42,7 @@ class Attachments(models.Model):
 
         storage.delete(path)
         thumb_storage.delete(thumb_path)
-        if not os.listdir(directory_path):
+        if os.path.isdir(directory_path) and not os.listdir(directory_path):
             os.rmdir(directory_path)
 
     issue = models.ForeignKey('Issues', models.DO_NOTHING,
@@ -192,6 +192,7 @@ class Issues(models.Model):
 
         for img in images_urls:
             if img and os.path.isfile(os.path.join(settings.MEDIA_ROOT, img)):
+                img_thumb = settings.MEDIA_URL + 'thumb-' + img
                 checked_img_urls.append(img)
 
         issue_obj = Issues.objects.filter(
